pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gtts>=2.5.4",
    "gunicorn>=23.0.0",
    "openai>=1.76.2",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "py-tgcalls>=2.1.1",
    "pydub>=0.25.1",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tgcrypto>=1.2.5",
]
